revno: 17908 message:   Made CategoryOptionGroupStore extend GenericNameableObjectStore
committer: Lars Helge Overland <larshelge@gmail.com>
branch nick: dhis2
timestamp: Wed 2015-01-07 18:57:12 +0100

+++ dhis-2/dhis-web/dhis-web-maintenance/dhis-web-maintenance-datadictionary/src/main/java/org/hisp/dhis/dd/action/categoryoptiongroup/ValidateCategoryOptionGroupAction.java
@@ -108,7 +108,7 @@ public class ValidateCategoryOptionGroupAction
         {
             CategoryOptionGroup match = dataElementCategoryService.getCategoryOptionGroupByName( name );
 
-            if ( match != null && (id == null || match.getId() != id.intValue()) )
+            if ( match != null && (id == null || match.getId() != id) )
             {
                 message = i18n.getString( "name_in_use" );
